change indentation style

+++ Src/CmTimer.h
@@ -1,10 +1,11 @@
 #pragma once
+
 #include "sys/time.h"
 
 class CmTimer
 {
 public:
-    CmTimer(CStr t):title(t) { is_started = false; gettimeofday(&start_clock,NULL); gettimeofday(&end_clock,NULL); n_starts = 0; }
+	CmTimer(CStr t):title(t) { is_started = false; gettimeofday(&start_clock,NULL); gettimeofday(&end_clock,NULL); n_starts = 0; }
 
 	~CmTimer(){	if (is_started) printf("CmTimer '%s' is started and is being destroyed.\n", title.c_str());	}
 
@@ -14,20 +15,20 @@ public:
 
 	inline bool Report();
 	inline bool StopAndReport() { Stop(); return Report(); }
-	inline float TimeInSeconds();
+	inline double TimeInSeconds();
 
 private:
 	CStr title;
 
 	bool is_started;
-    struct timeval start_clock, end_clock;
-    //clock_t start_clock;
-    //clock_t cumulative_clock;
+	struct timeval start_clock, end_clock;
+	//clock_t start_clock;
+	//clock_t cumulative_clock;
 	unsigned int n_starts;
 };
 
 /************************************************************************/
-/*                       Implementations                                */
+/*						 Implementations								*/
 /************************************************************************/
 
 void CmTimer::Start()
@@ -39,8 +40,8 @@ void CmTimer::Start()
 
 	is_started = true;
 	n_starts++;
-    //start_clock = clock();
-    gettimeofday(&start_clock,NULL);
+	//start_clock = clock();
+	gettimeofday(&start_clock,NULL);
 }
 
 void CmTimer::Stop()
@@ -49,8 +50,8 @@ void CmTimer::Stop()
 		printf("CmTimer '%s' is started. Nothing done\n", title.c_str());
 		return;
 	}
-    gettimeofday(&end_clock,NULL);
-    //cumulative_clock += clock() - start_clock;
+	gettimeofday(&end_clock,NULL);
+	//cumulative_clock += clock() - start_clock;
 	is_started = false;
 }
 
@@ -60,9 +61,9 @@ void CmTimer::Reset()
 		printf("CmTimer is started during reset request.\n Only reset cumulative time.\n");
 		return;
 	}
-    gettimeofday(&start_clock,NULL);
-    gettimeofday(&end_clock,NULL);
-    //cumulative_clock = 0;
+	gettimeofday(&start_clock,NULL);
+	gettimeofday(&end_clock,NULL);
+	//cumulative_clock = 0;
 }
 
 bool CmTimer::Report()
@@ -72,18 +73,18 @@ bool CmTimer::Report()
 		return false;
 	}
 
-	float timeUsed = TimeInSeconds();
+	double timeUsed = TimeInSeconds();
 	printf("[%s] CumuTime: %gs, #run: %d, AvgTime: %gs\n", title.c_str(), timeUsed, n_starts, timeUsed/n_starts);
 	return true;
 }
 
-float CmTimer::TimeInSeconds()
+double CmTimer::TimeInSeconds()
 {
 	if (is_started){
 		printf("CmTimer '%s' is started. Nothing done\n", title.c_str());
 		return 0;
 	}
-    return  double((end_clock.tv_sec  - start_clock.tv_sec) * 1000000u +
-             end_clock.tv_usec - start_clock.tv_usec) / 1.e6;
-    //return float(cumulative_clock) / CLOCKS_PER_SEC;
+	return double((end_clock.tv_sec	 - start_clock.tv_sec) * 1000000u +
+			 end_clock.tv_usec - start_clock.tv_usec) / 1.e6;
+	//return float(cumulative_clock) / CLOCKS_PER_SEC;
 }
